backend/requirements.txt
asgiref==3.4.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
cryptography==36.0.0
defusedxml==0.7.1
Django==3.2.9
django-3-jet==1.0.8
django-colorfield==0.5.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.2.1
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
packaging==21.3
pep8-naming==0.12.1
Pillow==8.4.0
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2021.3
reportlab==3.6.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
testfixtures==6.18.3
toml==0.10.2
tomli==2.0.0
uritemplate==4.1.1
urllib3==1.26.7
